`transformJSXComponentAtElementPathRecursively`
